save current modal build with seperate files
