+++ shift-scan/src/app/(routes)/admins/_pages/sidebar.tsx
@@ -4,22 +4,14 @@ import { Grids } from "@/components/(reusable)/grids";
 import { Holds } from "@/components/(reusable)/holds";
 import { Images } from "@/components/(reusable)/images";
 import { Texts } from "@/components/(reusable)/texts";
-import { signOut, useSession } from "next-auth/react";
-import { FormEvent, useEffect, useState } from "react";
+import { useSession } from "next-auth/react";
+import { useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
 import { usePathname } from "next/navigation";
 import { NModals } from "@/components/(reusable)/newmodals";
-import { Labels } from "@/components/(reusable)/labels";
-import { Inputs } from "@/components/(reusable)/inputs";
-import { Forms } from "@/components/(reusable)/forms";
-import { Contents } from "@/components/(reusable)/contents";
-import { Titles } from "@/components/(reusable)/titles";
-import { setUserLanguage, setUserPassword } from "@/actions/userActions";
-import { hash } from "bcryptjs";
-import { useTranslations } from "next-intl";
-import { useNotification } from "@/app/context/NotificationContext";
-import { Selects } from "@/components/(reusable)/selects";
-import { setLocale } from "@/actions/cookieActions";
+import PasswordModal from "./sidebar/PasswordModal";
+import LanguageModal from "./sidebar/LanguageModal";
+import SignOutModal from "./sidebar/SignOutModal";
 
 const Sidebar = () => {
   const pathname = usePathname();
@@ -28,166 +20,20 @@ const Sidebar = () => {
   const [username, setUsername] = useState<string>("");
   const { data: session } = useSession();
   const permission = session?.user.permission;
-  const userId = session?.user.id;
   const router = useRouter();
-  const { setNotification } = useNotification();
 
   const [isOpenChangePassword, setIsOpenChangePassword] = useState(false);
   const [isOpenLanguageSelector, setIsOpenLanguageSelector] = useState(false);
   const [isOpenSignOut, setIsOpenSignOut] = useState(false);
 
-  // change password modal props
-  const t = useTranslations("Hamburger");
-  const [showBanner, setShowBanner] = useState(false);
-  const [bannerMessage, setBannerMessage] = useState("");
-  const [eightChar, setEightChar] = useState(false);
-  const [oneNumber, setOneNumber] = useState(false);
-  const [oneSymbol, setOneSymbol] = useState(false);
-
-  const [viewSecret1, setViewSecret1] = useState(false);
-  const [viewSecret2, setViewSecret2] = useState(false);
-
-  const [newPassword, setNewPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
-
-  const viewPasscode1 = () => {
-    setViewSecret1(!viewSecret1);
-  };
-  const viewPasscode2 = () => {
-    setViewSecret2(!viewSecret2);
-  };
-
-  useEffect(() => {
-    if (showBanner) {
-      const timer = setTimeout(() => {
-        setShowBanner(false);
-      }, 8000); // Banner disappears after 5 seconds
-
-      return () => clearTimeout(timer); // Clear the timeout if the component unmounts
-    }
-  }, [showBanner]);
-
-  const handleSubmit = async (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-
-    if (newPassword.length === 0) {
-      setBannerMessage("Invalid. New Password cannot be empty.");
-      setShowBanner(true);
-      return;
-    }
-
-    if (confirmPassword.length === 0) {
-      setBannerMessage("Invalid. Confirm Password cannot be empty.");
-      setShowBanner(true);
-      return;
-    }
-
-    if (!validatePassword(newPassword)) {
-      setBannerMessage(
-        "Invalid. Password must be at least 8 characters long, contain 1 number, and 1 symbol."
-      );
-      setShowBanner(true);
-      return;
-    }
-
-    if (newPassword !== confirmPassword) {
-      setBannerMessage("Invalid. Passwords do not match!");
-      setShowBanner(true);
-      return;
-    }
-
-    const formData = new FormData(event.currentTarget);
-    const hashed = await hash(newPassword, 10);
-    formData.append("id", userId ?? "");
-    formData.append("password", hashed);
-
-    if (userId === "") {
-      setBannerMessage("Invalid. User ID cannot be empty.");
-      setShowBanner(true);
-      return;
-    }
-    try {
-      await setUserPassword(formData);
-      setIsOpenChangePassword(false);
-      setNotification("Password was successfully updated.", "success");
-    } catch (error) {
-      console.error("Error updating password:", error);
-      setNotification("Password was successfully updated.", "error");
-    }
-  };
-
-  const validatePassword = (password: string) => {
-    const minLength = 8;
-    const hasNumber = /\d/;
-    const hasSymbol = /[!@#$%^&*(),.?":{}|<>]/;
-
-    return (
-      password.length >= minLength &&
-      hasNumber.test(password) &&
-      hasSymbol.test(password)
-    );
-  };
-
-  const handlePasswordChange = (password: string) => {
-    setEightChar(password.length >= 8);
-    setOneNumber(/\d/.test(password));
-    setOneSymbol(/[!@#$%^&*(),.?":{}|<>]/.test(password));
-  };
-  const PasswordCriteria = ({
-    passed,
-    label,
-  }: {
-    passed: boolean;
-    label: string;
-  }) => (
-    <Holds position="row" className="space-x-2">
-      <Holds
-        background={passed ? "green" : "red"}
-        className="w-1 rounded-full my-auto"
-      ></Holds>
-      <Texts size="p6">{label}</Texts>
-    </Holds>
-  );
-  // end of change password modal
-
-  // language selector modal
-  const [language, setLanguage] = useState("en");
-  useEffect(() => {
-    const fetchLanguage = async () => {
-      try {
-        const response = await fetch(`/api/getSettings`);
-        if (response.ok) {
-          const data = await response.json();
-          setLanguage(data.language);
-        } else {
-          console.error("Error fetching language:", response.statusText);
-        }
-      } catch (error) {
-        console.error("Error fetching language:", error);
-      }
-    };
-    fetchLanguage();
-  }, [userId]);
-
-  const handleLanguageChange = async (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    const formData = new FormData(e.target as HTMLFormElement);
-    formData.append("id", userId ?? "");
-    const response = await setUserLanguage(formData);
-    if (response === "en") {
-      await setLocale(false);
-    } else {
-      await setLocale(true);
-    }
-  };
-
-  // end of language selector modal
   useEffect(() => {
     setUsername(`${session?.user.firstName} ${session?.user.lastName}`);
   }, [session?.user.firstName, session?.user.lastName]);
   return (
     <>
-      {/* If the side bar is closed it will show the mini menu */}
+      {/*----------------------------------------------------------------------------------------------------------*/}
+      {/*------------------------------- Displays the full extended sidebar ---------------------------------------*/}
+      {/*----------------------------------------------------------------------------------------------------------*/}
       {isOpen ? (
         <Holds background={"white"} className=" h-full w-[20em] ml-3 ">
           <Grids rows={"10"} gap={"5"}>
@@ -338,6 +184,8 @@ const Sidebar = () => {
               position={"row"}
               className="row-start-9 row-end-11 h-full gap-3 px-4"
             >
+              {/*----------------------------------------------------------------------------------------------------------*/}
+              {/*----------------------------------------------------------------------------------------------------------*/}
               {/* This is the password button */}
               <Buttons
                 background={"white"}
@@ -358,147 +206,12 @@ const Sidebar = () => {
                 isOpen={isOpenChangePassword}
                 handleClose={() => setIsOpenChangePassword(false)}
               >
-                <Holds background={"white"} className=" h-full px-4 ">
-                  <Forms
-                    onSubmit={handleSubmit}
-                    className="h-full flex flex-col items-center justify-between"
-                  >
-                    {showBanner && (
-                      <Holds
-                        background="red"
-                        position="absolute"
-                        size="full"
-                        className="rounded-none"
-                      >
-                        <Texts size="p6">{bannerMessage}</Texts>
-                      </Holds>
-                    )}
-
-                    {/* Start of grid container */}
-                    <Grids rows={"6"} gap={"5"} className="h-full">
-                      {/* New password section */}
-
-                      <Holds background="white" className="row-span-2 h-full">
-                        <Contents width="section">
-                          <Holds className="my-auto w-full">
-                            <Holds position="row" className="">
-                              <Labels htmlFor="new-password">
-                                {t("NewPassword")}
-                              </Labels>
-                              <Images
-                                titleImg={
-                                  viewSecret1 ? "/eye.svg" : "/eye-slash.svg"
-                                }
-                                titleImgAlt="eye"
-                                background="none"
-                                size="10"
-                                onClick={viewPasscode1}
-                              />
-                            </Holds>
-                            <Inputs
-                              type={viewSecret1 ? "text" : "password"}
-                              id="new-password"
-                              value={newPassword}
-                              onChange={(e) => {
-                                handlePasswordChange(e.target.value);
-                                setNewPassword(e.target.value);
-                              }}
-                            />
-                          </Holds>
-                        </Contents>
-                      </Holds>
-
-                      {/* Confirm password section */}
-                      <Holds className="row-span-4 h-full" background="white">
-                        <Contents width="section">
-                          <Holds className="my-auto w-full">
-                            <Holds position="row" className="h-full">
-                              <Labels htmlFor="confirm-password">
-                                {t("ConfirmPassword")}
-                              </Labels>
-                              <Images
-                                titleImg={
-                                  viewSecret2 ? "/eye.svg" : "/eye-slash.svg"
-                                }
-                                titleImgAlt="eye"
-                                background="none"
-                                size="10"
-                                onClick={viewPasscode2}
-                              />
-                            </Holds>
-                            <Inputs
-                              type={viewSecret2 ? "text" : "password"}
-                              id="confirm-password"
-                              value={confirmPassword}
-                              onChange={(e) =>
-                                setConfirmPassword(e.target.value)
-                              }
-                            />
-                          </Holds>
-                        </Contents>
-                      </Holds>
-                      <Holds
-                        background={"darkBlue"}
-                        className="row-span-1 h-full"
-                      >
-                        <Texts position="left" text={"white"} size="p4">
-                          Password Strength:
-                        </Texts>
-                        <Holds
-                          background="white"
-                          className="rounded-xl h-full "
-                        >
-                          <Contents width={"section"}>
-                            <Holds position="row" className="my-auto">
-                              <PasswordCriteria
-                                passed={oneNumber}
-                                label="123"
-                              />
-                              <PasswordCriteria
-                                passed={oneSymbol}
-                                label="Symbol"
-                              />
-                              <PasswordCriteria
-                                passed={eightChar}
-                                label="(8) Length"
-                              />
-                            </Holds>
-                          </Contents>
-                        </Holds>
-                      </Holds>
-
-                      {/* Submit button section */}
-                      <Holds className="row-span-2 h-full">
-                        <Holds className="my-auto flex justify-between">
-                          <Contents width="section" className="flex gap-4">
-                            {/* Submit Button */}
-                            <Buttons
-                              background="green"
-                              type="submit"
-                              className="py-2"
-                            >
-                              <Titles size="h4">{t("ChangePassword")}</Titles>
-                            </Buttons>
-
-                            {/* Cancel Button */}
-                            <Buttons
-                              background={"lightBlue"}
-                              type="button" // Prevents triggering form submission
-                              className="py-2"
-                              onClick={() => {
-                                // Logic to close the modal or reset form
-                                setIsOpenChangePassword(false);
-                              }}
-                            >
-                              <Titles size="h4">{t("Cancel")}</Titles>
-                            </Buttons>
-                          </Contents>
-                        </Holds>
-                      </Holds>
-                    </Grids>
-                  </Forms>
-                </Holds>
+                <PasswordModal
+                  setIsOpenChangePassword={() => setIsOpenChangePassword(false)}
+                />
               </NModals>
+              {/*----------------------------------------------------------------------------------------------------------*/}
+              {/*----------------------------------------------------------------------------------------------------------*/}
               {/* This is the language button */}
               <Buttons
                 background={"white"}
@@ -519,51 +232,14 @@ const Sidebar = () => {
                 isOpen={isOpenLanguageSelector}
                 handleClose={() => setIsOpenLanguageSelector(false)}
               >
-                <Holds background={"white"} className=" h-full w-full p-4">
-                  <Forms
-                    onSubmit={handleLanguageChange}
-                    className="h-full w-full"
-                  >
-                    <Texts size={"p4"} className="mb-4">
-                      Select A Language
-                    </Texts>
-                    <Holds className="my-auto h-1/3">
-                      <Selects
-                        id="language"
-                        name="language"
-                        className="w-full"
-                        value={language}
-                        onChange={(e) => {
-                          setLanguage(e.target.value);
-                        }}
-                      >
-                        <option value="en">English</option>
-                        <option value="es">Spanish</option>
-                      </Selects>
-                    </Holds>
-                    <Holds className="flex justify-between gap-5">
-                      <Buttons
-                        background={"green"}
-                        type="submit"
-                        className="py-2"
-                      >
-                        <Titles size="h4">{t("Save")}</Titles>
-                      </Buttons>
-                      <Buttons
-                        background={"lightBlue"}
-                        type="button" // Prevents triggering form submission
-                        className="py-2"
-                        onClick={() => {
-                          // Logic to close the modal or reset form
-                          setIsOpenLanguageSelector(false);
-                        }}
-                      >
-                        <Titles size="h4">{t("Cancel")}</Titles>
-                      </Buttons>
-                    </Holds>
-                  </Forms>
-                </Holds>
+                <LanguageModal
+                  setIsOpenLanguageSelector={() =>
+                    setIsOpenLanguageSelector(false)
+                  }
+                />
               </NModals>
+              {/*----------------------------------------------------------------------------------------------------------*/}
+              {/*----------------------------------------------------------------------------------------------------------*/}
               {/* This is the sign out button */}
               <Buttons
                 background={"white"}
@@ -583,45 +259,20 @@ const Sidebar = () => {
                 isOpen={isOpenSignOut}
                 handleClose={() => setIsOpenSignOut(false)}
               >
-                <Holds background={"white"} className=" h-full p-4">
-                  <Holds className="my-auto h-1/3">
-                    <Texts size={"p4"} className="mb-4">
-                      Are you sure you want to sign out?
-                    </Texts>
-                  </Holds>
-                  <Holds className=" flex flex-row gap-10">
-                    <Buttons
-                      type="button"
-                      onClick={async () => {
-                        setIsOpenSignOut(false); // Close the modal
-                        await signOut({
-                          redirect: true,
-                          callbackUrl: "/signin", // Specify the redirection URL
-                        });
-                      }}
-                      className="close-btn"
-                      background={"green"}
-                      size={"full"}
-                    >
-                      <Titles size={"h3"}>Yes</Titles>
-                    </Buttons>
-                    <Buttons
-                      type="button"
-                      onClick={() => setIsOpenSignOut(false)}
-                      className="close-btn"
-                      background={"red"}
-                      size={"full"}
-                    >
-                      <Titles size={"h3"}>Cancel</Titles>
-                    </Buttons>
-                  </Holds>
-                </Holds>
+                <SignOutModal
+                  setIsOpenSignOut={() => setIsOpenSignOut(false)}
+                />
               </NModals>
+              {/*----------------------------------------------------------------------------------------------------------*/}
+              {/*----------------------------------------------------------------------------------------------------------*/}
             </Holds>
           </Grids>
         </Holds>
       ) : (
         <Holds background={"white"} className=" h-full w-[5em] ml-3 ">
+          {/*----------------------------------------------------------------------------------------------------------*/}
+          {/*------------------------------------- Displays the mini menu --------------------------------------------*/}
+          {/*----------------------------------------------------------------------------------------------------------*/}
           <Grids rows={"10"} gap={"5"} className="my-5">
             <Holds className="row-span-4 h-full">
               <Holds className="w-24">
